@@ -1,11 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.9" />
-    <PackageReference Include="Microsoft.PowerBI.Api" Version="2.5.0" />
-    <PackageReference Include="Microsoft.PowerBI.JavaScript" Version="2.7.1" />
+    <PackageReference Include="Microsoft.PowerBI.Api" Version="3.11.1" />
+    <PackageReference Include="Microsoft.PowerBI.JavaScript" Version="2.11.0" />
   </ItemGroup>
   <ItemGroup>
     <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.4" />
